+++ packages/core/src/constants/chainId.ts
@@ -5,6 +5,7 @@ import {
   Arbitrum,
   ArbitrumRinkeby,
   ArbitrumGoerli,
+  ArbitrumNova,
   Avalanche,
   AvalancheTestnet,
   Aurora,
@@ -61,6 +62,8 @@ import {
   KlaytnTestnet,
   Klaytn,
   ScrollAlphaTestnet,
+  LineaTestnet,
+  MantleTestnet,
 } from '../model'
 
 // rough alphabet order (put network from the same chain together)
@@ -72,6 +75,7 @@ export const DEFAULT_SUPPORTED_CHAINS = [
   Arbitrum,
   ArbitrumRinkeby,
   ArbitrumGoerli,
+  ArbitrumNova,
   Aurora,
   AuroraTestnet,
   Mainnet,
@@ -126,6 +130,8 @@ export const DEFAULT_SUPPORTED_CHAINS = [
   BaseGoerli,
   Base,
   ScrollAlphaTestnet,
+  LineaTestnet,
+  MantleTestnet,
 ]
 
 export enum ChainId {
@@ -190,4 +196,7 @@ export enum ChainId {
   BaseGoerli = 84531,
   Base = 8453,
   ScrollAlpha = 534353,
+  LineaTestnet = 59140,
+  ArbitrumNova = 42170,
+  MantleTestnet = 5001,
 }