+++ packages/connectors/wallet-connect-v2/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @usedapp/wallet-connect-v2-connector
 
+## 1.0.4
+
+### Patch Changes
+
+- 9b69b70: Add Readme
+
+## 1.0.3
+
+### Patch Changes
+
+- 819e1bf: Use @wallectconnect/modal instead of @web3modal/standalone
+
 ## 1.0.2
 
 ### Patch Changes